Same MAC Fixes #568
